tests/wpt/tests/css/css-grid/grid-model/fixed-width-intrinsic-width-should-exclude-scrollbar-001.html

<!DOCTYPE html>
<html>
<title>CSS Grid: intrinsic width of fixed-width grid items.</title>
<link rel="author" title="Sunil Ratnu" href="mailto:sunil.ratnu@samsung.com"/>
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#overflow"/>
<meta name="assert" content="This test ensures that scrollbar width is not accounted for fixed-width grid items' intrinsic width."/>
<link href="/css/support/grid.css" rel="stylesheet"/>
<style>
.grid {
    grid-template-columns: 400px 500px;
    grid-template-rows: 200px 300px;
}

.gridItemScrollOverflow {
    width: 50px;
    height: 50px;
    overflow: scroll;
}

</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<body onload="checkLayout('.grid')">

<div class="grid">
    <div class="firstRowFirstColumn gridItemScrollOverflow" data-expected-width="50" data-expected-height="50"></div>
    <div class="firstRowSecondColumn gridItemScrollOverflow" data-expected-width="50" data-expected-height="50"></div>
    <div class="secondRowFirstColumn gridItemScrollOverflow" data-expected-width="50" data-expected-height="50"></div>
    <!-- Grid item itself being a grid container-->
    <div class="grid secondRowSecondColumn gridItemScrollOverflow" data-expected-width="50" data-expected-height="50"></div>
</div>

</body>
</html>
